# Enable compile-time-perf for HIPCC Jenkins build