--- a/core/CMakeLists.txt
+++ b/core/CMakeLists.txt
@@ -11,4 +11,11 @@ KOKKOS_ADD_TEST_DIRECTORIES(perf_test)
 
 KOKKOS_SUBPACKAGE_POSTPROCESS()
 
-
+IF(Kokkos_ENABLE_TESTS)
+  ADD_TEST(
+    NAME                compile-time-perf
+    COMMAND             ${CMAKE_COMMAND}
+                        --build ${PROJECT_BINARY_DIR}
+                        --target analyze-kokkos-compile-time
+    WORKING_DIRECTORY   ${PROJECT_BINARY_DIR})
+ENDIF()