--- a/scripts/docker/Dockerfile.hipcc
+++ b/scripts/docker/Dockerfile.hipcc
@@ -5,8 +5,9 @@ RUN apt-get update && apt-get install -y \
         git \
         kmod \
         wget \
-        ccache \
         file \
+        ccache \
+        python3 \
         && \
     apt-get clean && \
     rm -rf /var/lib/apt/lists/*