Will generate a new test named `compile-time-perf` which will contain data such as this:

```console
Start testing: Mar 31 23:06 PDT
----------------------------------------------------------
1/47 Testing: compile-time-perf
1/47 Test: compile-time-perf
Command: "/opt/local/bin/cmake" "--build" "/Users/jrmadsen/devel/c++/kokkos/kokkos-nersc/build" "--target" "analyze-kokkos-compile-time"
Directory: /Users/jrmadsen/devel/c++/kokkos/kokkos-nersc/build
"compile-time-perf" start time: Mar 31 23:06 PDT
Output:
----------------------------------------------------------
[1/1] Analyzing compile times from .compile-time-perf-timem-output...

Common arguments to all commands:
    -DNDEBUG    
    -MD         
    -MF         
    -MT         
    -O2         
    -c          
    -g          
    -isysroot   
    -mmacosx-version-min=10.15
    -o          
    /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX11.1.sdk

wall_clock         110.250 sec  g++ core/unit_test/openmp/TestOpenMP_MathematicalFunctions.cpp
wall_clock         107.050 sec  g++ core/unit_test/serial/TestSerial_ViewLayoutStrideAssignment.cpp
wall_clock          99.402 sec  g++ core/unit_test/openmp/TestOpenMP_ViewLayoutStrideAssignment.cpp
wall_clock          94.917 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c05.cpp
wall_clock          91.709 sec  g++ core/unit_test/serial/TestSerial_ViewResize.cpp
wall_clock          86.736 sec  g++ core/unit_test/serial/TestSerial_SubView_c04.cpp
wall_clock          86.568 sec  g++ core/unit_test/serial/TestSerial_MathematicalFunctions.cpp
wall_clock          86.324 sec  g++ core/unit_test/serial/TestSerial_SubView_c05.cpp
wall_clock          85.799 sec  g++ core/unit_test/serial/TestSerial_Other.cpp
wall_clock          84.676 sec  g++ core/unit_test/openmp/TestOpenMP_Other.cpp
wall_clock          84.333 sec  g++ ../core/unit_test/default/TestDefaultDeviceTypeResize.cpp
wall_clock          83.826 sec  g++ core/unit_test/openmp/TestOpenMP_ViewResize.cpp
wall_clock          82.666 sec  g++ core/unit_test/serial/TestSerial_SubView_c06.cpp
wall_clock          81.882 sec  g++ containers/unit_tests/openmp/TestOpenMP_ScatterView.cpp
wall_clock          81.828 sec  g++ core/unit_test/openmp/TestOpenMP_DeepCopyAlignment.cpp
wall_clock          79.294 sec  g++ core/unit_test/openmp/TestOpenMP_ViewMapping_a.cpp
wall_clock          77.591 sec  g++ core/unit_test/serial/TestSerial_ViewMapping_a.cpp
wall_clock          77.089 sec  g++ core/unit_test/serial/TestSerial_SubView_c07.cpp
wall_clock          76.675 sec  g++ core/unit_test/openmp/TestOpenMP_LocalDeepCopy.cpp
wall_clock          75.281 sec  g++ core/unit_test/serial/TestSerial_DeepCopyAlignment.cpp
wall_clock          72.897 sec  g++ containers/unit_tests/serial/TestSerial_ScatterView.cpp
wall_clock          71.154 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c11.cpp
wall_clock          70.714 sec  g++ core/unit_test/serial/TestSerial_SubView_c11.cpp
wall_clock          70.589 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c10.cpp
wall_clock          70.587 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c12.cpp
wall_clock          70.131 sec  g++ core/unit_test/serial/TestSerial_SubView_c12.cpp
wall_clock          69.233 sec  g++ core/unit_test/serial/TestSerial_SubView_c10.cpp
wall_clock          69.153 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c06.cpp
wall_clock          69.146 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c04.cpp
wall_clock          66.478 sec  g++ core/unit_test/serial/TestSerial_LocalDeepCopy.cpp
wall_clock          65.956 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c08.cpp
wall_clock          65.318 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c09.cpp
wall_clock          65.234 sec  g++ core/unit_test/serial/TestSerial_SubView_c08.cpp
wall_clock          65.118 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c07.cpp
wall_clock          64.427 sec  g++ core/unit_test/serial/TestSerial_SubView_c09.cpp
wall_clock          61.698 sec  g++ core/unit_test/openmp/TestOpenMP_TeamBasic.cpp
wall_clock          61.485 sec  g++ core/unit_test/openmp/TestOpenMP_NumericTraits.cpp
wall_clock          57.292 sec  g++ core/unit_test/serial/TestSerial_SubView_a.cpp
wall_clock          52.324 sec  g++ core/unit_test/serial/TestSerial_ViewCopy_a.cpp
wall_clock          50.541 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_a.cpp
wall_clock          44.476 sec  g++ core/unit_test/serial/TestSerial_NumericTraits.cpp
wall_clock          41.076 sec  g++ ../algorithms/unit_tests/TestSerial.cpp 
wall_clock          38.937 sec  g++ core/unit_test/openmp/TestOpenMP_Atomics.cpp
wall_clock          36.946 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_c3.cpp
wall_clock          35.766 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_b3.cpp
wall_clock          35.641 sec  g++ core/unit_test/serial/TestSerial_Atomics.cpp
wall_clock          35.387 sec  g++ ../core/perf_test/test_atomic_minmax_simple.cpp
wall_clock          34.306 sec  g++ containers/unit_tests/openmp/TestOpenMP_DynViewAPI_generic.cpp
wall_clock          33.961 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_a3.cpp
wall_clock          33.767 sec  g++ containers/unit_tests/serial/TestSerial_DynViewAPI_generic.cpp
wall_clock          33.675 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_c2.cpp
wall_clock          32.992 sec  g++ core/unit_test/serial/TestSerial_Reducers_d.cpp
wall_clock          32.722 sec  g++ core/unit_test/openmp/TestOpenMP_Reducers_d.cpp
wall_clock          32.689 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_b2.cpp
wall_clock          32.605 sec  g++ core/unit_test/serial/TestSerial_ViewAPI_a.cpp
wall_clock          32.465 sec  g++ core/unit_test/openmp/TestOpenMP_MDRange_d.cpp
wall_clock          30.998 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_a2.cpp
wall_clock          30.748 sec  g++ core/unit_test/serial/TestSerial_MDRange_d.cpp
wall_clock          30.592 sec  g++ ../core/unit_test/serial/TestSerial_Task.cpp
wall_clock          29.860 sec  g++ core/unit_test/serial/TestSerial_ViewAPI_b.cpp
wall_clock          29.813 sec  g++ containers/unit_tests/serial/TestSerial_OffsetView.cpp
wall_clock          28.864 sec  g++ containers/unit_tests/openmp/TestOpenMP_DualView.cpp
wall_clock          28.838 sec  g++ containers/unit_tests/openmp/TestOpenMP_OffsetView.cpp
wall_clock          28.505 sec  g++ core/unit_test/serial/TestSerial_TeamScan.cpp
wall_clock          28.405 sec  g++ ../core/src/impl/Kokkos_Core.cpp        
wall_clock          28.292 sec  g++ containers/unit_tests/serial/TestSerial_DualView.cpp
wall_clock          28.246 sec  g++ core/unit_test/openmp/TestOpenMP_AtomicViews.cpp
wall_clock          27.729 sec  g++ core/unit_test/openmp/TestOpenMP_TeamScan.cpp
wall_clock          27.346 sec  g++ core/unit_test/serial/TestSerial_AtomicViews.cpp
wall_clock          27.106 sec  g++ core/unit_test/openmp/TestOpenMP_ViewAPI_b.cpp
wall_clock          26.963 sec  g++ core/unit_test/serial/TestSerial_ViewAPI_e.cpp
wall_clock          26.067 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_b.cpp
wall_clock          25.775 sec  g++ core/unit_test/serial/TestSerial_SubView_b.cpp
wall_clock          25.341 sec  g++ core/unit_test/serial/TestSerial_MDRange_c.cpp
wall_clock          25.113 sec  g++ core/unit_test/openmp/TestOpenMP_MDRange_c.cpp
wall_clock          24.721 sec  g++ core/unit_test/openmp/TestOpenMP_MDRange_e.cpp
wall_clock          24.305 sec  g++ core/unit_test/openmp/TestOpenMP_ViewAPI_e.cpp
wall_clock          24.155 sec  g++ core/unit_test/serial/TestSerial_MDRange_b.cpp
wall_clock          24.065 sec  g++ core/unit_test/serial/TestSerial_MDRange_e.cpp
wall_clock          23.905 sec  g++ core/unit_test/openmp/TestOpenMP_Reducers_a.cpp
wall_clock          23.878 sec  g++ core/unit_test/openmp/TestOpenMP_MDRange_b.cpp
wall_clock          23.698 sec  g++ ../core/unit_test/openmp/TestOpenMP_Task.cpp
wall_clock          23.369 sec  g++ core/unit_test/openmp/TestOpenMP_Reducers_b.cpp
wall_clock          23.217 sec  g++ core/unit_test/serial/TestSerial_MDRange_a.cpp
wall_clock          23.085 sec  g++ core/unit_test/openmp/TestOpenMP_ViewCopy_a.cpp
wall_clock          22.836 sec  g++ ../algorithms/unit_tests/TestOpenMP_Random.cpp
wall_clock          22.809 sec  g++ core/unit_test/openmp/TestOpenMP_MDRange_a.cpp
wall_clock          22.195 sec  g++ core/unit_test/serial/TestSerial_Reducers_a.cpp
wall_clock          22.128 sec  g++ core/unit_test/serial/TestSerial_Reducers_b.cpp
wall_clock          21.843 sec  g++ core/unit_test/serial/TestSerial_TeamBasic.cpp
wall_clock          21.820 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_c1.cpp
wall_clock          21.283 sec  g++ core/unit_test/openmp/TestOpenMP_RangePolicyRequire.cpp
wall_clock          21.083 sec  g++ ../core/perf_test/test_atomic.cpp       
wall_clock          20.831 sec  g++ core/unit_test/openmp/TestOpenMP_RangePolicy.cpp
wall_clock          20.642 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_b1.cpp
wall_clock          20.527 sec  g++ ../core/unit_test/default/TestDefaultDeviceType_a1.cpp
wall_clock          20.396 sec  g++ core/unit_test/openmp/TestOpenMP_ViewAPI_a.cpp
wall_clock          20.224 sec  g++ ../core/perf_test/PerfTest_ViewResize_123.cpp
wall_clock          19.503 sec  g++ ../tpls/gtest/gtest/gtest-all.cc        
wall_clock          19.456 sec  g++ containers/unit_tests/openmp/TestOpenMP_StaticCrsGraph.cpp
wall_clock          19.350 sec  g++ core/unit_test/openmp/TestOpenMP_Reducers_c.cpp
wall_clock          18.987 sec  g++ core/unit_test/serial/TestSerial_ViewAPI_d.cpp
wall_clock          18.968 sec  g++ core/unit_test/generated/Test08_deep_copy_OPENMP.cpp
wall_clock          18.813 sec  g++ core/unit_test/serial/TestSerial_Reducers_c.cpp
wall_clock          18.801 sec  g++ ../core/unit_test/default/TestDefaultDeviceType.cpp
wall_clock          18.606 sec  g++ core/unit_test/serial/TestSerial_RangePolicyRequire.cpp
wall_clock          18.605 sec  g++ core/unit_test/openmp/TestOpenMP_Init.cpp
wall_clock          18.523 sec  g++ containers/unit_tests/serial/TestSerial_StaticCrsGraph.cpp
wall_clock          18.492 sec  g++ core/unit_test/serial/TestSerial_RangePolicy.cpp
wall_clock          18.041 sec  g++ ../tpls/gtest/gtest/gtest-all.cc        
wall_clock          17.968 sec  g++ ../core/perf_test/PerfTest_ViewResize_45.cpp
wall_clock          17.484 sec  g++ core/unit_test/openmp/TestOpenMP_Reductions.cpp
wall_clock          17.365 sec  g++ core/unit_test/openmp/TestOpenMP_TeamScratch.cpp
wall_clock          16.535 sec  g++ ../core/unit_test/openmp/TestOpenMP_Graph.cpp
wall_clock          16.506 sec  g++ core/unit_test/serial/TestSerial_Init.cpp
wall_clock          16.310 sec  g++ core/unit_test/serial/TestSerial_Reductions.cpp
wall_clock          15.989 sec  g++ core/unit_test/openmp/TestOpenMP_ViewAPI_d.cpp
wall_clock          15.775 sec  g++ ../core/unit_test/serial/TestSerial_Graph.cpp
wall_clock          15.639 sec  g++ core/unit_test/serial/TestSerial_SubView_c02.cpp
wall_clock          15.604 sec  g++ containers/unit_tests/openmp/TestOpenMP_UnorderedMap.cpp
wall_clock          15.585 sec  g++ ../algorithms/unit_tests/TestOpenMP_SortDynamicView.cpp
wall_clock          15.534 sec  g++ core/unit_test/serial/TestSerial_SubView_c01.cpp
wall_clock          15.515 sec  g++ core/unit_test/openmp/TestOpenMP_TeamTeamSize.cpp
wall_clock          15.443 sec  g++ core/unit_test/serial/TestSerial_SubView_c03.cpp
wall_clock          15.165 sec  g++ ../core/perf_test/PerfTest_ViewFill_45.cpp
wall_clock          15.154 sec  g++ ../core/perf_test/PerfTest_ViewFill_123.cpp
wall_clock          15.103 sec  g++ core/unit_test/serial/TestSerial_ViewCopy_b.cpp
wall_clock          14.972 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c02.cpp
wall_clock          14.871 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c03.cpp
wall_clock          14.842 sec  g++ core/unit_test/serial/TestSerial_TeamScratch.cpp
wall_clock          14.797 sec  g++ core/unit_test/openmp/TestOpenMP_TeamVectorRange.cpp
wall_clock          14.736 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c01.cpp
wall_clock          14.562 sec  g++ core/unit_test/generated/Test08_deep_copy_SERIAL.cpp
wall_clock          14.428 sec  g++ ../core/perf_test/PerfTest_ViewResize_8.cpp
wall_clock          14.396 sec  g++ core/unit_test/serial/TestSerial_TeamVectorRange.cpp
wall_clock          14.278 sec  g++ ../containers/performance_tests/TestOpenMP.cpp
wall_clock          14.267 sec  g++ core/unit_test/serial/TestSerial_ViewAPI_c.cpp
wall_clock          14.263 sec  g++ ../core/perf_test/PerfTest_ViewResize_6.cpp
wall_clock          14.042 sec  g++ ../core/perf_test/PerfTest_ViewResize_7.cpp
wall_clock          14.016 sec  g++ containers/unit_tests/openmp/TestOpenMP_ErrorReporter.cpp
wall_clock          14.006 sec  g++ core/unit_test/openmp/TestOpenMP_UniqueToken.cpp
wall_clock          13.910 sec  g++ core/unit_test/serial/TestSerial_UniqueToken.cpp
wall_clock          13.774 sec  g++ containers/unit_tests/serial/TestSerial_ErrorReporter.cpp
wall_clock          13.498 sec  g++ core/unit_test/serial/TestSerial_WorkGraph.cpp
wall_clock          13.228 sec  g++ core/unit_test/openmp/TestOpenMP_WorkGraph.cpp
wall_clock          13.099 sec  g++ containers/unit_tests/openmp/TestOpenMP_Vector.cpp
wall_clock          13.055 sec  g++ ../algorithms/unit_tests/TestOpenMP_Sort3D.cpp
wall_clock          13.049 sec  g++ ../algorithms/unit_tests/TestOpenMP_Sort1D.cpp
wall_clock          13.049 sec  g++ core/unit_test/openmp/TestOpenMP_Complex.cpp
wall_clock          12.766 sec  g++ containers/unit_tests/serial/TestSerial_UnorderedMap.cpp
wall_clock          12.464 sec  g++ core/unit_test/openmp/TestOpenMP_ViewCopy_b.cpp
wall_clock          12.328 sec  g++ ../algorithms/unit_tests/TestOpenMP.cpp 
wall_clock          12.317 sec  g++ core/unit_test/openmp/TestOpenMP_ViewAPI_c.cpp
wall_clock          12.163 sec  g++ ../core/perf_test/PerfTest_ExecSpacePartitioning.cpp
wall_clock          12.087 sec  g++ core/unit_test/serial/TestSerial_Complex.cpp
wall_clock          12.015 sec  g++ core/unit_test/openmp/TestOpenMP_TeamReductionScan.cpp
wall_clock          11.906 sec  g++ core/unit_test/openmp/TestOpenMP_Crs.cpp
wall_clock          11.756 sec  g++ core/unit_test/openmp/TestOpenMP_Reductions_DeviceView.cpp
wall_clock          11.595 sec  g++ ../core/perf_test/PerfTest_ViewCopy_d45.cpp
wall_clock          11.484 sec  g++ ../core/perf_test/PerfTest_ViewCopy_d123.cpp
wall_clock          11.472 sec  g++ ../core/perf_test/PerfTest_ViewCopy_c45.cpp
wall_clock          11.440 sec  g++ containers/unit_tests/serial/TestSerial_Vector.cpp
wall_clock          11.426 sec  g++ ../core/perf_test/PerfTest_ViewFill_8.cpp
wall_clock          11.408 sec  g++ core/unit_test/serial/TestSerial_Reductions_DeviceView.cpp
wall_clock          11.338 sec  g++ ../core/perf_test/PerfTest_ViewCopy_c123.cpp
wall_clock          11.151 sec  g++ ../core/perf_test/PerfTest_ViewFill_7.cpp
wall_clock          10.950 sec  g++ core/unit_test/serial/TestSerial_TeamTeamSize.cpp
wall_clock          10.839 sec  g++ ../core/perf_test/PerfTest_ViewFill_6.cpp
wall_clock          10.791 sec  g++ ../core/perf_test/PerfTest_ViewAllocate.cpp
wall_clock          10.655 sec  g++ core/unit_test/serial/TestSerial_ViewMapping_b.cpp
wall_clock          10.639 sec  g++ containers/unit_tests/serial/TestSerial_DynViewAPI_rank12345.cpp
wall_clock          10.635 sec  g++ core/unit_test/openmp/TestOpenMP_AtomicOperations_unsignedint.cpp
wall_clock          10.590 sec  g++ core/unit_test/openmp/TestOpenMP_AtomicOperations_longint.cpp
wall_clock          10.568 sec  g++ core/unit_test/openmp/TestOpenMP_AtomicOperations_int.cpp
wall_clock          10.506 sec  g++ core/unit_test/openmp/TestOpenMP_AtomicOperations_unsignedlongint.cpp
wall_clock          10.460 sec  g++ containers/unit_tests/openmp/TestOpenMP_ViewCtorPropEmbeddedDim.cpp
wall_clock          10.448 sec  g++ core/unit_test/openmp/TestOpenMP_AtomicOperations_longlongint.cpp
wall_clock          10.397 sec  g++ ../core/perf_test/PerfTest_ViewCopy_b45.cpp
wall_clock          10.386 sec  g++ core/unit_test/generated/Test06_ParallelFor_MDRangePolicy_OPENMP.cpp
wall_clock          10.327 sec  g++ core/unit_test/serial/TestSerial_Crs.cpp
wall_clock          10.303 sec  g++ containers/unit_tests/serial/TestSerial_DynamicView.cpp
wall_clock          10.274 sec  g++ ../core/perf_test/PerfTest_ViewCopy_a123.cpp
wall_clock          10.272 sec  g++ core/unit_test/serial/TestSerial_ViewMapping_subview.cpp
wall_clock          10.261 sec  g++ ../core/perf_test/PerfTest_ViewCopy_b123.cpp
wall_clock          10.094 sec  g++ ../core/perf_test/PerfTestGramSchmidt.cpp
wall_clock          10.048 sec  g++ ../core/perf_test/PerfTest_ViewCopy_a45.cpp
wall_clock           9.878 sec  g++ core/unit_test/serial/TestSerial_TeamReductionScan.cpp
wall_clock           9.391 sec  g++ containers/unit_tests/serial/TestSerial_Bitset.cpp
wall_clock           9.346 sec  g++ ../core/perf_test/test_taskdag.cpp      
wall_clock           9.328 sec  g++ core/unit_test/openmp/TestOpenMP_ViewMapping_b.cpp
wall_clock           9.304 sec  g++ core/unit_test/serial/TestSerial_SubView_c14.cpp
wall_clock           9.303 sec  g++ containers/unit_tests/openmp/TestOpenMP_DynamicView.cpp
wall_clock           9.264 sec  g++ core/unit_test/generated/Test04_ParallelFor_RangePolicy_OPENMP.cpp
wall_clock           9.240 sec  g++ ../core/unit_test/default/TestDefaultDeviceTypeInit_13.cpp
wall_clock           9.194 sec  g++ ../core/perf_test/PerfTest_CustomReduction.cpp
wall_clock           9.083 sec  g++ core/unit_test/openmp/TestOpenMP_ViewMapping_subview.cpp
wall_clock           8.996 sec  g++ core/unit_test/generated/Test14_MDRangeReduce_OPENMP.cpp
wall_clock           8.942 sec  g++ containers/unit_tests/openmp/TestOpenMP_Bitset.cpp
wall_clock           8.789 sec  g++ containers/unit_tests/openmp/TestOpenMP_DynViewAPI_rank12345.cpp
wall_clock           8.727 sec  g++ ../core/unit_test/default/TestDefaultDeviceTypeInit_11.cpp
wall_clock           8.716 sec  g++ ../core/unit_test/default/TestDefaultDeviceTypeInit_12.cpp
wall_clock           8.712 sec  g++ core/unit_test/generated/Test10_HierarchicalBasics_OPENMP.cpp
wall_clock           8.692 sec  g++ ../core/perf_test/PerfTest_ViewCopy_d6.cpp
wall_clock           8.689 sec  g++ core/unit_test/serial/TestSerial_ViewOfClass.cpp
wall_clock           8.643 sec  g++ ../core/perf_test/PerfTest_ViewCopy_c6.cpp
wall_clock           8.439 sec  g++ core/unit_test/generated/Test03b_MemorySpace_free_OPENMP.cpp
wall_clock           8.399 sec  g++ ../core/perf_test/PerfTest_ViewCopy_d8.cpp
wall_clock           8.389 sec  g++ core/unit_test/openmp/TestOpenMP_SubView_c14.cpp
wall_clock           8.381 sec  g++ core/unit_test/generated/Test03a_MemorySpace_malloc_OPENMP.cpp
wall_clock           8.363 sec  g++ ../core/perf_test/PerfTest_ViewCopy_c7.cpp
wall_clock           8.357 sec  g++ ../core/perf_test/PerfTest_ViewCopy_c8.cpp
wall_clock           8.278 sec  g++ ../core/unit_test/default/TestDefaultDeviceTypeInit_15.cpp
wall_clock           8.277 sec  g++ ../core/perf_test/PerfTest_ViewCopy_d7.cpp
wall_clock           8.257 sec  g++ containers/unit_tests/serial/TestSerial_ViewCtorPropEmbeddedDim.cpp
wall_clock           8.238 sec  g++ core/unit_test/serial/TestSerial_AtomicOperations_longlongint.cpp
wall_clock           8.191 sec  g++ core/unit_test/serial/TestSerial_AtomicOperations_unsignedlongint.cpp
wall_clock           8.107 sec  g++ core/unit_test/openmp/TestOpenMP_SharedAlloc.cpp
…