@masterleinad 2021-04-08
```
Scanning dependencies of target analyze-kokkos-compile-time
Analyzing compile times from .compile-time-perf-timem-output...
Traceback (most recent call last):
  File "/usr/local/bin/compile-time-perf-analyzer", line 415, in <module>
    args = parser.parse_intermixed_args(_iargs)
AttributeError: 'ArgumentParser' object has no attribute 'parse_intermixed_args'
CMakeFiles/analyze-kokkos-compile-time.dir/build.make:60: recipe for target 'CMakeFiles/analyze-kokkos-compile-time' failed
make[3]: *** [CMakeFiles/analyze-kokkos-compile-time] Error 1
CMakeFiles/Makefile2:598: recipe for target 'CMakeFiles/analyze-kokkos-compile-time.dir/all' failed
make[2]: *** [CMakeFiles/analyze-kokkos-compile-time.dir/all] Error 2
CMakeFiles/Makefile2:605: recipe for target 'CMakeFiles/analyze-kokkos-compile-time.dir/rule' failed
make[1]: *** [CMakeFiles/analyze-kokkos-compile-time.dir/rule] Error 2
Makefile:178: recipe for target 'analyze-kokkos-compile-time' failed
```