@jrmadsen 2021-04-08
@Rombur that commit was more of an experiment on my end to see what needed to be fixed. Useful too since I didn't realize that the parse intermixed args in pythons argparse was relatively new.