@jrmadsen 2021-04-21
> Please split the individual changes (changes to RULE_LAUCNH_COMPILER, changes to CMake installation, using compile-time-perf) up into multiple pull requests.

What? These all go together. I'll back out the changes to every Jenkins build but I'm not going to create multiple pull requests for things that go together. That's like asking me to create separate pull requests for each new command line argument.