@DavidPoliakoff 2021-04-21
@jrmadsen @masterleinad : I actually don't think you guys are too far apart. @jrmadsen, I'd recommend you just motivate why all this is in one place a bit for Daniel (I'm guessing you're using rule_launch_compile to exec you wrapper script for timing or something?)