@jrmadsen 2021-04-21
I tweaked the RULE_LAUNCH_COMPILE because the compile-time-perf also uses it and the way the kokkos rule launch stuff is designed, it always needs to be last because it depends on finding a match between one of the parameters it holds and the first argument of the command.

The cmake modifications were done at @masterleinad recommendation but because the gpg stuff is so buggy, every retest ended up with a failure somewhere so then I'd retest and I'd get a different container fail to build. This issue has been around for a while. So since we are the Kokkos development team and not the CMake development team and we have to write old-style cmake anyway bc tribits, we don't need to do their continuous integration for them. Thus, I kept a test for our minimum CMake version and replaced all the buggy SHA stuff with a much more stable implementation in order to actually get all the containers to build properly so that they could actually be tested.