@masterleinad 2021-04-21
If the issue with `RULE_LAUNCH_COMPILER` really only is triggered here, I don't have strong objections to leaving it here (although it very likely gets merged faster if it's pulled into a separate pull request that can also demonstrate the issue elsewhere).
I feel stronger about changing the way `CMake` is pulled/installed. The pull request also works without this change so it should be separate IMHO. Note that I'm not opposed to that change in the first place.