@jrmadsen 2021-06-09
@arghdos You may be interested in the [Jenkins artifacts](https://cloud.cees.ornl.gov/jenkins-ci/blue/organizations/jenkins/pykokkos-base/detail/main/2/artifacts) as part of the [pykokkos-base](https://github.com/kokkos/pykokkos-base) CI. 

The [CUDA-pykokkos-base-compile-time.txt](https://cloud.cees.ornl.gov/jenkins-ci/job/pykokkos-base/job/main/2/artifact/build/CUDA-pykokkos-base-compile-time.txt) and [HIP-pykokkos-base-compile-time.txt](https://cloud.cees.ornl.gov/jenkins-ci/job/pykokkos-base/job/main/2/artifact/build/HIP-pykokkos-base-compile-time.txt) are basically a head-to-head comparison of NVCC vs. HIPCC compile times for the same set of View/DynRankView template instantiations, e.g.:

```
View<..., Kokkos::CudaSpace, ...>
...
View<..., Kokkos::CudaHostPinnedSpace, ...>
...
```

vs.
 
```
View<..., Kokkos::HIPSpace, ...>
...
View<..., Kokkos::HIPHostPinnedSpace, ...>
...
``` 

HIPCC comes out way ahead of NVCC.

## CUDA-pykokkos-base-compile-time.txt

```console
Common arguments to all commands:
    -DKOKKOS_DEPENDENCE
    -DNDEBUG    
    -O3         
    -arch=sm_70 
    -fPIC       
    -o          
    /usr/bin/c++
    /usr/local/bin/nvcc_wrapper

wall_clock         333.595 sec  kokkos_launch_compiler src/variants/concrete_view_uint64_left_managed.cpp
wall_clock         330.190 sec  kokkos_launch_compiler src/variants/concrete_view_uint16_left_managed.cpp
wall_clock         326.530 sec  kokkos_launch_compiler src/variants/concrete_view_int16_right_managed.cpp
...
peak_rss        4624.792 MB   kokkos_launch_compiler src/variants/concrete_view_float32_right_managed.cpp
peak_rss        4624.716 MB   kokkos_launch_compiler src/variants/concrete_view_float64_right_managed.cpp
peak_rss        4624.688 MB   kokkos_launch_compiler src/variants/concrete_view_float32_left_managed.cpp
...
```

## HIP-pykokkos-base-compile-time.txt

```console
Common arguments to all commands:
    -DKOKKOS_DEPENDENCE
    -DNDEBUG    
    -O3         
    -fPIC       
    -fno-gpu-rdc
    -o          
    /opt/rocm/bin/hipcc
    /usr/bin/c++

wall_clock          69.022 sec  kokkos_launch_compiler src/variants/concrete_view_uint16_right_managed.cpp
wall_clock          68.962 sec  kokkos_launch_compiler src/variants/concrete_view_uint16_left_managed.cpp
wall_clock          68.529 sec  kokkos_launch_compiler src/variants/concrete_view_int16_left_managed.cpp
...
peak_rss        1258.120 MB   kokkos_launch_compiler src/variants/concrete_view_uint16_right_managed.cpp
peak_rss        1256.768 MB   kokkos_launch_compiler src/variants/concrete_view_int64_right_managed.cpp
peak_rss        1256.304 MB   kokkos_launch_compiler src/variants/concrete_view_int32_right_managed.cpp
...
```