@masterleinad 2022-06-21
I think this would be nice to have to use in a performance regression framework once that exists.